test(stylex-transform): put the operand property beside the operators
Issue 04 asked for the three-operator property "beside the existing
`nodes/tests/logical_expression_tests.rs`" and it landed in the shape suite
two directories up instead. It is a property of `&&`, `||` and `??` -- that
a right operand they cannot fold refuses rather than aborting -- so it belongs
with them; the shape suite keeps the catalogue of shapes that must refuse
wherever they sit, and each file now names the other.

The move needs the parse-and-evaluate helpers in one place, which they were not:
both suites had grown their own `evaluate_source`, `parse_expr`,
`assert_deopts` and two `assert_folds_to_*`. Two suites asserting "this
refuses" through separately maintained helpers is how one of them comes to
accept a refusal carrying no reason, so they collapse into
`tests/source_evaluation.rs` and all three callers share it.
