+++ crates/stylex-transform/src/shared/utils/js/evaluate/tests/member_length_tests.rs
@@ -14,165 +14,7 @@
 //! each answer something else on non-ASCII input, so the unicode cases are the
 //! load-bearing half of this file.
 
-use super::*;
-use stylex_structures::stylex_options::StyleXOptions;
-use swc_core::{
-  common::{FileName, GLOBALS, Globals, SourceMap, sync::Lrc},
-  ecma::parser::{EsSyntax, Parser, StringInput, Syntax, lexer::Lexer},
-};
-
-/// Parses one expression and evaluates it. Panics propagate: a refusal that
-/// aborts the build instead of deopting is the defect
-/// [#1265](https://github.com/Dwlad90/stylex-swc-plugin/issues/1265) reported,
-/// and every refusal below sits one operand away from that seam.
-fn evaluate_source(source: &str) -> Box<EvaluateResult> {
-  let expr = parse_expr(source);
-  let globals = Globals::new();
-
-  GLOBALS.set(&globals, || {
-    let mut traversal_state = StateManager::new(StyleXOptions::default());
-    let fns = FunctionMap::default();
-
-    evaluate(&expr, &mut traversal_state, &fns)
-  })
-}
-
-fn parse_expr(source: &str) -> Expr {
-  let source_map: Lrc<SourceMap> = Default::default();
-  let source_file = source_map.new_source_file(FileName::Anon.into(), source.to_string());
-
-  let lexer = Lexer::new(
-    Syntax::Es(EsSyntax {
-      jsx: true,
-      ..Default::default()
-    }),
-    Default::default(),
-    StringInput::from(&*source_file),
-    None,
-  );
-
-  match Parser::new_from(lexer).parse_expr() {
-    Ok(expr) => *expr,
-    Err(error) => panic!("failed to parse `{}`: {:?}", source, error),
-  }
-}
-
-/// Asserts the source folds to a number. Spelled as an exact count rather than
-/// "some number", because the whole failure mode this file guards is a
-/// confident answer that is not the right one.
-#[track_caller]
-fn assert_folds_to_number(source: &str, expected: f64) {
-  let result = evaluate_source(source);
-
-  assert!(
-    result.confident,
-    "expected `{}` to fold, got a deopt: {:?}",
-    source, result.reason
-  );
-
-  match result.value {
-    Some(EvaluateResultValue::Expr(Expr::Lit(Lit::Num(num)))) => {
-      assert_eq!(num.value, expected, "wrong folded number for `{}`", source)
-    },
-    other => panic!("expected `{}` to fold to a number, got {:?}", source, other),
-  }
-}
-
-/// Asserts the source folds to a string. Used for the fallback a missing
-/// property lets through, where the interesting half is that the fold happened
-/// at all.
-#[track_caller]
-fn assert_folds_to_string(source: &str, expected: &str) {
-  let result = evaluate_source(source);
-
-  assert!(
-    result.confident,
-    "expected `{}` to fold, got a deopt: {:?}",
-    source, result.reason
-  );
-
-  match result.value {
-    Some(EvaluateResultValue::Expr(Expr::Lit(Lit::Str(strng)))) => assert_eq!(
-      convert_atom_to_string(&strng.value),
-      expected,
-      "wrong folded string for `{}`",
-      source
-    ),
-    other => panic!("expected `{}` to fold to a string, got {:?}", source, other),
-  }
-}
-
-/// Asserts the source folds to `undefined` — a value the evaluator is confident
-/// about, not a refusal. That answer exists so a declaration folds its fallback
-/// instead of reaching the runtime.
-#[track_caller]
-fn assert_folds_to_undefined(source: &str) {
-  let result = evaluate_source(source);
-
-  assert!(
-    result.confident,
-    "expected `{}` to fold to undefined, got a deopt: {:?}",
-    source, result.reason
-  );
-
-  match result.value {
-    Some(EvaluateResultValue::Expr(Expr::Ident(ident))) => assert_eq!(
-      ident.sym.as_ref(),
-      "undefined",
-      "wrong folded identifier for `{}`",
-      source
-    ),
-    other => panic!(
-      "expected `{}` to fold to undefined, got {:?}",
-      source, other
-    ),
-  }
-}
-
-/// Asserts the source refuses, as a deopt carrying a reason rather than as an
-/// abort. The reason has to be there: `stylex.create()` turns it into the
-/// author-facing diagnostic.
-#[track_caller]
-fn assert_deopts(source: &str) {
-  let result = evaluate_source(source);
-
-  assert!(
-    !result.confident,
-    "expected `{}` to refuse to fold, got {:?}",
-    source, result.value
-  );
-
-  assert!(
-    result.reason.is_some(),
-    "expected `{}` to record a deopt reason",
-    source
-  );
-}
-
-/// Asserts the refusal names the property that could not be read. The node kind
-/// is the half an author can already see; which property was asked for is the
-/// half that says why the declaration will not fold.
-#[track_caller]
-fn assert_deopt_names_property(source: &str, property: &str) {
-  let result = evaluate_source(source);
-
-  assert!(
-    !result.confident,
-    "expected `{}` to refuse to fold, got {:?}",
-    source, result.value
-  );
-
-  match result.reason {
-    Some(reason) => assert!(
-      reason.contains(property),
-      "expected the refusal of `{}` to name `{}`, got {:?}",
-      source,
-      property,
-      reason
-    ),
-    None => panic!("expected `{}` to record a deopt reason", source),
-  }
-}
+use super::source_evaluation::*;
 
 // ==================== the fold this file is about ====================
